Move F2 adapter helpers out of recipe internals (#524)
Relocate check_adapter_base_path, rewrite_adapter_config_base_path,
stash_adapter_files (+ STALE_BASE_PATH_TAG) from custom_recipes/
custom_recipe_utils.py to a new tools/torchtune/adapter_utils.py.

These helpers are cross-tool — inspect setup/submit import them too —
so they don't belong in recipe internals. custom_recipe_utils.py now
holds only validate_epochs_to_save (F1, still recipe-owned). Repoints
post_finetune.py and the two inspect-side importers; tests follow.

Also addresses review feedback: drop #465 references from
post_finetune.py docstring and trim over-explaining comments from
finetune_template.slurm.
