+++ src/tools/inspect/setup_inspect.py
@@ -19,7 +19,7 @@ import warnings
 import yaml
 from pathlib import Path
 
-from cruijff_kit.tools.torchtune.custom_recipes.custom_recipe_utils import (
+from cruijff_kit.tools.torchtune.adapter_utils import (
     check_adapter_base_path,
 )
 from cruijff_kit.tools.torchtune.model_configs import MODEL_CONFIGS
